backend/requirements.txt
# API Framework
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.17
pydantic==2.9.2
click==8.1.7

# Database
peewee==3.17.8
psycopg2-binary==2.9.9

# Data Processing & Machine Learning
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.3.0

# Deep Learning & Vision
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
transformers==4.34.0
tokenizers==0.14.1  # Updated to be compatible with transformers 4.34.0

# Weather API & HTTP
openmeteo-requests==1.3.0
requests-cache==1.1.1
retry-requests==2.0.0
openmeteo-sdk==1.18.0
requests==2.31.0

# Image Processing
Pillow==10.0.0
matplotlib==3.7.2

# Utilities
python-dateutil>=2.8.2
six>=1.16.0
typing-extensions>=4.5.0
urllib3>=2.0.0
setuptools>=69.0.3
wheel>=0.42.0
charset-normalizer>=3.1.0
idna>=3.4
pyparsing>=3.0.9
contourpy>=1.1.0
cycler>=0.11.0
fonttools>=4.47.2
kiwisolver>=1.4.5
packaging>=23.2
zipp>=3.17.0
importlib-metadata>=6.8.0
safetensors>=0.3.1
regex>=2023.6.3
tqdm>=4.66.1
sympy>=1.12
networkx>=3.1
filelock>=3.12.2
fsspec>=2023.6.0
jinja2>=3.1.2
markupsafe>=2.1.3

# Optional but recommended for better performance
ninja>=1.11.1
